Continued creating new worklog-time-entry
Removed package time-interval package; will only use package
local-time-duration.

Added some methods to determine whether daylight savings time
is in effect, and the appropriate offset to use when converting
a worklog datetime into a timestamp.

Continue working on creating a new class; it is not ready yet
but the package compiles without error.